break up new crate into smaller modules
