+++ harness/src/fuzz/firedancer.rs
@@ -19,7 +19,7 @@ use {
         metadata::Metadata as FuzzMetadata,
         Fixture as FuzzFixture,
     },
-    mollusk_svm_result::result::InstructionResult,
+    mollusk_svm_result::InstructionResult,
     solana_account::Account,
     solana_compute_budget::compute_budget::ComputeBudget,
     solana_instruction::{error::InstructionError, AccountMeta, Instruction},
